fix(web): resolve CI lint + test failures
- Move draft-send-in-flight gate functions to @/lib/draft-send-gate.ts
  (shared top-level module) to fix cross-domain import lint error — chat
  domain hooks were importing from conversations domain

- Update post-chat-message test to flush microtask queue before asserting
  workspace writes, since persistPreChatOnboardingProfile is now
  fire-and-forget
