+++ apps/web/src/domains/chat/chat-layout.tsx
@@ -24,10 +24,10 @@ import type { AssistantContextValue } from "@/components/layout/assistant-contex
 import { useConversationStore } from "@/domains/conversations/conversation-store.js";
 import {
   chatContextQueryKey,
-  hasDraftSendInFlight,
   useConversationGroupsQuery,
   useConversationListQuery,
 } from "@/domains/conversations/conversation-queries.js";
+import { hasDraftSendInFlight } from "@/lib/draft-send-gate.js";
 import { useAttentionTracking } from "@/domains/conversations/use-attention-tracking.js";
 import { useConversationActions } from "@/domains/conversations/use-conversation-actions.js";
 import { useConversationGroupActions } from "@/domains/conversations/use-conversation-group-actions.js";
