+++ apps/web/src/generated
@@ -1,0 +1,1 @@
+/Users/jasonzhou/Vellum/vellum-assistant/apps/web/src/generated
